refactor(config): Simplify URL fetching policy implementation
Moves the detailed `UrlFetchPolicy` dataclass and associated environment variable parsing helpers from `bibra/config.py` to `net_security.py` where they belong better
