+++ bibra/net_security.py
@@ -28,16 +28,141 @@ Fetch components:
 import asyncio
 import ipaddress
 import logging
+import os
 import re
 import socket
 import urllib.parse
+from dataclasses import dataclass
 from typing import Protocol
 
 import httpx2
 
-from bibra.config import URL_FETCH_DIRECT, UrlFetchPolicy
-
 logger = logging.getLogger(__name__)
+
+
+# Sentinel value for BIBRA_URL_PROXY meaning "fetch directly, with the
+# full in-app URL validation". Any other non-blank value is treated as a
+# proxy URL; a blank/unset value means URL fetching is refused entirely.
+URL_FETCH_DIRECT = "direct"
+
+# Defaults for the URL fetch policy (BIBRA_URL_* env vars).
+DEFAULT_URL_SCHEMES: list[str] = ["https"]
+DEFAULT_URL_CONTENT_TYPES: list[str] = ["application/pdf"]
+DEFAULT_URL_MAX_BYTES: int = 50 * 1024 * 1024  # 50 MiB
+DEFAULT_URL_TIMEOUT: int = 30  # seconds
+DEFAULT_URL_MAX_REDIRECTS: int = 5
+DEFAULT_URL_ALLOW_IP_HOSTS: bool = False
+
+
+@dataclass(frozen=True)
+class UrlFetchPolicy:
+    """Policy governing how BIBRA fetches user-supplied URLs (SSRF hardening).
+
+    Attributes:
+        proxy: The egress proxy URL, the literal sentinel "direct" (see
+            URL_FETCH_DIRECT) for direct egress, or None to refuse fetching.
+        schemes: Allowed URL schemes (e.g. ["https"]).
+        content_types: Allowed response content types (MIME, no parameters).
+        max_bytes: Hard cap on total downloaded bytes.
+        timeout: Whole seconds for connect/read/write/pool timeouts.
+        max_redirects: Maximum number of redirect hops; every hop is
+            re-validated against the policy.
+        allow_ip_hosts: Whether URLs whose host is an IP literal are allowed.
+    """
+
+    proxy: str | None
+    schemes: tuple[str, ...]
+    content_types: tuple[str, ...]
+    max_bytes: int
+    timeout: int
+    max_redirects: int
+    allow_ip_hosts: bool
+
+    @property
+    def proxy_required(self) -> bool:
+        """True when URL fetching must be refused (no proxy or direct set)."""
+        return self.proxy is None
+
+
+def _parse_list_env(name: str, default: list[str]) -> tuple[str, ...]:
+    """Parse a comma-separated env var into a tuple of stripped strings.
+
+    Falls back to the default when the variable is unset or blank.
+    """
+    raw = os.environ.get(name, "")
+    items = [item.strip().lower() for item in raw.split(",") if item.strip()]
+    return tuple(items) if items else tuple(item.lower() for item in default)
+
+
+def _parse_int_env(name: str, default: int) -> int:
+    """Parse a positive-integer env var, falling back to the default."""
+    raw = os.environ.get(name, "").strip()
+    if not raw:
+        return default
+    try:
+        value = int(raw)
+    except ValueError:
+        logger.debug("Invalid %s=%r; using default %d", name, raw, default)
+        return default
+    return value if value > 0 else default
+
+
+def _parse_bool_env(name: str, default: bool) -> bool:
+    """Parse a boolean env var (1/true/yes/on), falling back to the default."""
+    raw = os.environ.get(name, "").strip().lower()
+    if not raw:
+        return default
+    return raw in {"1", "true", "yes", "on"}
+
+
+def get_url_proxy() -> str | None:
+    """Return the BIBRA_URL_PROXY environment variable value.
+
+    Blank or whitespace-only values are normalized to None, so that
+    callers can safely pass the result to httpx proxy arguments.
+
+    Returns:
+        The proxy URL string, or None if not set or blank.
+    """
+    proxy = os.environ.get("BIBRA_URL_PROXY")
+    return proxy.strip() if proxy and proxy.strip() else None
+
+
+def load_url_fetch_policy(cli_fallback: bool = False) -> UrlFetchPolicy:
+    """Build a UrlFetchPolicy from the BIBRA_URL_* environment variables.
+
+    Semantics of BIBRA_URL_PROXY:
+      - unset/blank: URL fetching is refused (proxy_required is True)
+      - "direct": direct egress is allowed, subject to full in-app validation
+      - anything else: used as the egress proxy URL
+
+    With ``cli_fallback=True`` (the CLI's intent), an unset/blank
+    BIBRA_URL_PROXY is treated as "direct" instead of refusing: the CLI
+    is a local tool and falls back to direct egress with full in-app
+    validation, while the REST API keeps refusing by default.
+
+    Invalid numeric/boolean values are logged and replaced by their
+    defaults, so that a misconfigured setting never crashes the app.
+    """
+    proxy = get_url_proxy()
+    if proxy is None and cli_fallback:
+        proxy = URL_FETCH_DIRECT
+    return UrlFetchPolicy(
+        proxy=proxy,
+        schemes=_parse_list_env("BIBRA_URL_SCHEMES", DEFAULT_URL_SCHEMES),
+        content_types=_parse_list_env(
+            "BIBRA_URL_CONTENT_TYPES", DEFAULT_URL_CONTENT_TYPES
+        ),
+        max_bytes=_parse_int_env("BIBRA_URL_MAX_BYTES", DEFAULT_URL_MAX_BYTES),
+        timeout=_parse_int_env("BIBRA_URL_TIMEOUT", DEFAULT_URL_TIMEOUT),
+        max_redirects=_parse_int_env(
+            "BIBRA_URL_MAX_REDIRECTS", DEFAULT_URL_MAX_REDIRECTS
+        ),
+        allow_ip_hosts=_parse_bool_env(
+            "BIBRA_URL_ALLOW_IP_HOSTS", DEFAULT_URL_ALLOW_IP_HOSTS
+        ),
+    )
+
 
 #: Read the response body in chunks no larger than this while enforcing the
 #: size cap, so we can abort mid-stream without buffering unbounded data.
